+++ src/winconfig/default/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dbosoft/winconfig-default
 
+## 0.3.0
+
+### Minor Changes
+
+- 08b4953: Move configuration of SSH server feature in Windows to separate gene
+
 ## 0.2.0
 
 ### Minor Changes